refactor: move ReconnectingOverlay to molecules/dialog
Extract the overlay from App.kt into the shared dialog components package without behavior changes.
